+++ js/network.js
@@ -1,4 +1,4 @@
-let adsArrey;
+
 function getDataFromServer(onSuccess, onError){
   return fetch('https://25.javascript.pages.academy/keksobooking/data')
     .then((response) => {
@@ -10,13 +10,10 @@ function getDataFromServer(onSuccess, onError){
     })
     .then((response) => response.json())
     .then((data) => {
-      //console.log('Результат', data);
       onSuccess(data);
-      adsArrey = data;
     })
     .catch(() => {
-      onError();//показать сообщение об ошибке (сваять) , если при загрузке данных с сервера произошла ошибка запроса
-      //console.log(error);
+      onError();
     });
 }
 
@@ -45,4 +42,4 @@ function sendFormData (form, onSuccess, onError) {
     });
 }
 
-export {getDataFromServer, sendFormData, adsArrey};
+export {getDataFromServer, sendFormData };
