+++ js/util.js
@@ -1,4 +1,5 @@
-//имя_функции(от, до);
+const ALERT_SHOW_TIME = 7000;
+
 // Результат: целое число из диапазона "от...до"
 function getRandomInteger(min, max) {
   if(max > min && min >= 0){
@@ -52,4 +53,28 @@ const debounce = (callback, timeoutDelay) => {
   };
 };
 
-export {getRandomFractionalNumber, getRandomInteger, getArrayRandomElement, getArrayRandomElements, boolean, isEscapeKey, isNumberIntervalIncluded, debounce};
+//показать сообщение об ошибке, если при загрузке данных с сервера произошла ошибка запроса
+function errorMessage  () {
+  const alertContainer = document.createElement('div');
+  alertContainer.style.zIndex = '1000';
+  alertContainer.style.position = 'fixed';
+  alertContainer.style.width = '500px';
+  alertContainer.style.top = '100px';
+  alertContainer.style.left = '50%';
+  alertContainer.style.transform = 'translateX(-50%)';
+  alertContainer.style.padding = '50px 30px';
+  alertContainer.style.fontSize = '20px';
+  alertContainer.style.lineHeight = '26px';
+  alertContainer.style.textAlign = 'center';
+  alertContainer.style.backgroundColor = '#fff';
+  alertContainer.style.borderRadius = '10px';
+  alertContainer.textContent = 'При загрузке данных с сервера произошла ошибка, но вы можете опубликовать свое объявление :)';
+
+  document.body.append(alertContainer);
+
+  setTimeout(() => {
+    alertContainer.remove();
+  }, ALERT_SHOW_TIME);
+}
+
+export {getRandomFractionalNumber, getRandomInteger, getArrayRandomElement, getArrayRandomElements, boolean, isEscapeKey, isNumberIntervalIncluded, debounce, errorMessage};
